refactor([issue-4901]): move auth gate into services
